Trip/Template Trip: Add Destination adding with confirmation
This is similar to the behavior in regards to adding new Clients. It's
worth noting that adding a new Client will use the Pick-Up address as
their home address. So that is a case where we don't automatically check
"Add Pick-Up address to Destinations". The user can still explicitly
check that option if desired.