Report: Add support for per-driver reports
These reports can be accessed by going to the Drivers page, clicking
on the Edit dropdown next to one of the drivers, and then clicking on
'View report'.

There are some differences with how the driver report compares to the
standard report. These are listed in an alert at the top of the page.