Schedule: Add 'Driver Status' trip format
This format can be used to indicate driver availability on the schedule.
It works like Activities, execpt there is a field for the driver, as
well as the boolean for their availability. The reasoning for adding
this as a separate feature is:

- The desire for the color shading of driver availability to match the
  existing driver colors
- More consistency in displaying available vs not available

Unlike Activities and Trips, Driver Statuses can't be marked as
canceled. I don't think this would make much logical sense, as they
don't describe any specific happening.

Bugs fixed:
- Fixed the display of start/end time of non-Trips on the print version
  of the schedule
- Fixed missing targets in the Event Log filter