Trip: Only use "active" shift when getting previous mileage data
This fixes a bug when the same vehicle was used for more than one shift
on a single day. The problem being that it would try to merge the short
hand trip mileages with the end mileage of the latest completed shift.
So as a fix, we only use trips that have a driver matching
a non-completed shift. Otherwise, we just use the last shift mileage
with no attempt to merge trips.