Shift/Trip: Attempt to provide better suggestions for mileage typos
The following suggestion methods have been added:
- Remove the first character
- "Shift" the characters over if a decimal point was potentially missed

But instead of providing multiple suggestions, we rank them by score and
provide the best one. The score in this case is difference in mileage
versus the reference mileage.

Also defined now is an upper threshold for mileage. In other words,
we'll offer suggestions if it looks like the mileage might be too high.
This has been set to a slightly arbritrary 250 miles. The highest
real-world mileage from existing data is around 80 miles. So the chances
of a false positive here are low.

Because there is now an upper threshold, we may end up with no
reasonable suggestions. In such a case, we now simply display the option
to clear the field and try again. I think this is a better option than
presenting the user with a poor (and most likely inaccurate) suggestion.