Report: By default, don't show daily data when report contains >31 days
When trying to view a report that contains *more* than 31 days, a button
is shown that will override this behavior and show all data.

The purpose of this addition is to cut down on the amount of stuff that
has to be rendered in the template. This speeds up loading quite a bit
when, for example, looking at a 12-month report. In such scenarios, it
usually isn't desired to view the daily data, anyway.