Report: Use shift end mileage when checking for trip mileage validity
For example, consider the following shift mileage:
19945.5 -> 20019.5

And the following trip mileage entered as-is:
999.9 -> 002.9

Before the changes in this commit, the trip end mileage would evaluate
to 19002.9 and result in an error. Now, it will attempt to merge the
trip mileage with the shift end mileage to see if the result makes
sense. In this case, the trip end mileage would evaluate to 20002.9,
which is in the bounds of the shift mileage range.