Client & Destination: Support updating trips when duplicates exist
This fixes the issue where updating a client or destination would skip
the "Update Trips" page if there was a duplicate. Now, we correctly
prompt to fix dupes *after* trips have been updated. If no trips need to
be updated, it will proceed to the dupe fix page as normal.

We also now bold the original entry in the dupe list, which can help
tell them apart.