Vehicle Status: Redesigned layout
The purpose of this is the display the information in a more organized
way *by vehicle* as opposed to by data type. It's also designed to look
better on mobile, where drivers might be using this page.

The issues and pretrips can still be viewed in the old way, but are now
on separate pages. One minor "bug" is that creating or editing an issue
from the Issue Tracker page will return the user to the Vehicle Status
page instead.